feat : add the payment logic to make the payment related to email add the welcome page in the URL/api make the logic to add the succes payment to subscriptions table

+++ database/migrations/2026_05_26_145739_add_plain_id_column_to_payment_table.php
@@ -8,8 +8,13 @@ return new class extends Migration
 {
     public function up(): void
     {
+        if (Schema::hasColumn('payments', 'plan_id')) {
+            Schema::table('payments', function (Blueprint $table) {
+                $table->dropColumn('plan_id');
+            });
+        }
+
         Schema::table('payments', function (Blueprint $table) {
-            $table->dropColumn('plan_id');
             $table->unsignedBigInteger('plan_id')->nullable()->after('service_id'); // ← nullable
 
             $table->foreign('plan_id')->references('id')->on('service_plans')->onDelete('cascade');
@@ -18,9 +23,11 @@ return new class extends Migration
 
     public function down(): void
     {
-        Schema::table('payments', function (Blueprint $table) {
-            $table->dropForeign(['plan_id']);
-            $table->dropColumn('plan_id');
-        });
+        if (Schema::hasColumn('payments', 'plan_id')) {
+            Schema::table('payments', function (Blueprint $table) {
+                $table->dropForeign(['plan_id']);
+                $table->dropColumn('plan_id');
+            });
+        }
     }
 };
